Change 3
You don't need to delete the vector. It is created in the stack and after execution of the function it will be deleted in a neat way. Deleting blocks in vectors manually especially by starting from the beginning (data.begin()) rather than end is a very expensive action! That's the reason those two functions( Depth histogram Collisions histogram) which delete the data vector at the end take 0.2 seconds to print out the histograms but 10 seconds to go back to the main menu.